requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
pgvector>=0.4.0
sentence-transformers>=2.6.1
transformers>=4.40.0
torch>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
